I have verified that the values computed by rama here match the beta16 scores computed by R3 for ubiquitin;
The trick in comparing these scores vs R3's scores is that
because R3 is treating RamaPrePro as a two-body energy, half
of the score is assigned to residue i, and the other half
is assigned to residue i+1. R3 also applies the beta_nov16.wts
weight of 0.5 to the rama term.