Determine rama table indices from residue properties
RamaScoreGraph now reads from a new class, ResidueProperties, at
construction (as well as from the PolymericBonds::upper) to
determine which table each residue should use to calculate its
energy.

score_support now as a single-dispatch for ResidueProperties'
factory method when constructing one from a PackedStructure.

Currently, the table-index deduction logic is CPU-locked and
is written with python for loops and lots of python operations
(instead of some magical C-function from a library like pandas
that does the substring matching I want). It would be nice
if this setup phase could be written with a magical library,
though!