Fix the numeric deriv check unit test
The previous version (and the version still present in
test_total_gradcheck.py) ignores the incoming coordinates
and so there deriv check is not actually running.

Enabling the real deriv check to happen results in a failure
unless the eps step is upped to 1e-3 and the rtol and atol are
similarly increased.