Add Ramachandran tables representation to database-layer (from PR #79)
This commit adds the tables, in json, for the Ramachandran potential.
These tables are converted into bspline coefficients for later
interpolation; retriving the coefficient tables is memoized so that only
one copy of each set of tables (you can imagine different json files
defining different tables) will reside in GPU memory at a time.

The json file also defines the rules for which table to use for
which residue in the Pose based on its residue properties and its
upper-neighbor's properties. This will be used in conjuction with
PR #117, which defines those property hierarchies.