Add non-Ramachandran score components that Rama needs (from PR #79)
This PR introduces three scoring mix-ins needed to evaluate Rama,
and that will be similarly used for Dunbrack:
 - AlphaAABackboneTorsionProvider
 - PolymericBonds, and
 - ResidueProperties

The AlphaAABackboneTorsionProvider measures the torsions for
particular residues in the system; the indices of the atoms
defining those torsions are provided by the PackedSystem
(or whatever else defines the score graph).

The PolymericBonds class gives the residue-to-residue indices
for chemical bonds, so that residue i can know the index of
the residue(s) it is chemically bonded to (which isn't always
i+1). This is needed by Rama so that it can know what the
upper residue is for residue i to decide between the normal-
and pre-proline tables. These indices are provided by the
PackedSystem.

The ResidueProperties class holds arrays of property lists
for each of the residues in the ScoreGraph. These properties
are copied out of the ResidueTypes by the PackedSystem.

Finally, this PR adds access to double-precision representation
of the system's coordinates. This is needed as there is a
startling loss of precision when going from dihedrals to
coordinates then back to dihedrals.

This PR depends on PRs #116 and #117