Add Ramachandran score graph (from PR #79)
The culmination of PRs #116, #117, #118, and #119: the evaluation
of the Ramachandran score in the ScoreGraph.

The RamaIntraScore is created by the RamaScoreGraph class during
score evaluation, pulling the torsions and table ids from the
"target", as these will be shared. It has a "total_rama" method
returning the sum of the per-residue ramachandran energies,
which are themselves calculated by b-spline interpolation of
the tables that live in the database.