Reformat the rama.json file for a more compact representation
Instead of storing a long list of (phi,psi,prob,energy) tuples,
the tables are now represented as actual tables, where the phi/psi
value can be inferred from the indices in the table.

The file is formatted so that the first value at phi/psi (-180,-180)
appears at the BOTTOM left corner so that it has the visual appearance
of the ramachandran plot with (0,0) in the center, the beta region
in the upper left, and the helical region in the lower/middle left.