Read Ramachandran tables from binary zarr archive
This commit adds code for properly converting the starting ramachandran
database file (from R3) to a binary file that is readable by tmol.
It then reads that file in to the data structures in

   tmol/database/scoring/rama.py

This file has been significantly simplified, eschewing the cattr.structure
methods in favor of direct component construction, and eliminating the
read-from-json-to-intermediate --> write-to-binary and
read-from-binary-to-intermediate --> construct completed data structure
steps.

There are assuredly linting fixes that need to be made.