Add zarr to core.requirements.txt; clean linting errors
NOTE: the environments/README.md suggests that pip-installable modules
belong in core.dependencies.txt, however, when I put zarr into
core.dependencies.txt, then ./dev_setup produced an error saying
zarr was not a recognized module. I'm not sure if the documentation
is backwards, or if something else is going on with that module
that I'm not understanding. I'm 99% certain I used pip to install
zarr locally the first time I tried to use it, as opposed to
running a conda command.