Add Ramachandran database layer (#118)
This commit adds a binary representation of the Ramachandran database to tmol as well as the converter script to read Rosetta's original .txt file and convert it into the binary format. The Ramachandran database includes the tables for the 40 variations (broken down by 20 AAs and yes/no is the next residue a proline). The mapping that informs which table should be used for which residues in the structure is mediated by residue-properties: the database layer states which table should be used for a pair of properties (one for residue i, one for residue i+1), and this is converted into a pair of regexes for matching against the properties of the residues in the structure.