Dunbrack appears to be working but does not generate R3's scores
It looks like the difference in interpolation schemes is the source
of the difference. The chi means are different between R3 and tmol,
as well as the chi stdevs. Furthermore, the interpolated -ln(prob)s are
also different.

On the whole, however, the scores are all in the ballpark of the R3
scores, +/- 0.5 kcal/mol.

This is not very satisfying!