Save before restructing to get autograd::gradcheck to pass
gradcheck is complaining that I have not taken into account
the dV_dE that is passed to the DunbrackFunc.forward function.
The reason I do this is because it feels faster to just multiply
the dE_ddihes and ddihe_dxyzs on the forward pass. OK. So I guess
I'd better split this multiplication out for later.