Fix proline scores for dunbrack
OK, proline wasn't getting scored correctly because the rotamer
library lists the two rotamers with indices "1 1 1" and "2 1 1",
but based on the chi angles, they would be classified using
the g+/t/g- logic as "1 3 1" and "3 1 3".  So I'm saying for proline:
if you're in "1 3 1" or "3 1 3" score yourself as if you were
in the same bin as "1 1 1" or "2 1 1".

This actually will probably work as a good solution to the non-existant
rotamer problem for lys / arg (but probably not needed on the database
level).