switch undefined dihedral convention
Now if any atoms are undefined for a dihedral, then all four atom incides
should be set to a negative value.

-1 is the sentinel used for an undefined phi (setting it to a neutral value)
-2 is the sentinel used for an undefined psi (setting it to a neutral value)

negative values broadly are used to prevent dihedrals from being computed
that are undefined and for their derivatives to be accumulated; if neither
-1 nor -2 are given, then the diheral value that is input will be preserved
(whatever that happened to be; perhaps garbage).