Pre-convert Lists of Tensors to TCollections
TCollections spent a lot of time transfering a small vector of
memory pointers to the GPU before every function call; better
to just send this vector of memory pointers to the GPU once and
then hang on to the created TCollection.

Dunbrack forward() evaluation is now down to 475 us. Still achingly
slow.