Convert list of Tensors to a C++ TensorCollection object
By converting a list of Tensors to a C++ object and then returning
that object back to python, then subsequent calls to C++ functions
with pybind no longer require the somewhat-expensive construction
of the TensorCollection object. In particular, the TensorCollection
object needs to send an array of memory addresses to the GPU in its
constructor, and this transfer to the GPU at every energy evaluation
is terribly time consuming.

Updating the Rama potential to use this create-tensor-collection-once
paradigm.

One other change with this PR: TViews now be directly cast to
TensorAccessors; slices of TViews produce TensorAccessors, so this
change means that a function like the bspline interpolation
that wants an XD array can be fed input from an XD TView or a
slide of a (X+1)D TView. Bspline interpolation functions now
take a TensorAccessor.