Dunbrack score (#196)
Dunbrack score graph.

The Dunbrack library is organized into four sets of tables:
 - the rotameric rotamer-probability tables, 1 per rotamer per amino acid, 2D
 - the rotamer-mean tables, 1 per chi per rotamer per amino acid, 2D
 - the rotamer-stdev tables, 1 per chi per rotamer per amino acid, 2D
 - the semirotameric rotamer-probability tables, 1 per rotamer per amino acid, 3D

Each set of tables is packed into a single tensor of 1-higher dimension. In order to support the eventual creation of tables at a lower resolution, each of these 1-higher-dimensional tables is supplemented with a size and stride data (with mean & stdev tables sharing this data, since their sizes and strides are the same). This avoids the use of TCollections, which are slow to cast.

The evaluation of the energy occurs in 5 kernels:
 1. measure the dihedrals (bb & chi)
 2. determine rotamer indices based on dihedrals
 3. compute chi-deviation penalty for rotatmeric chi
 4. compute -ln(prob) energy for rotameric aas
 5. compute -ln(prob) energy for semi-rotameric aas

Steps 3-5 could reasonably occur in parallel.

One thing to note with this implementation. All of the interpolation tables are put into an (N+1)D tensor, and then slices of this tensor are passed into the interpolation function. To slice a tensor, I have to provide tensor sizes and tensor strides. E.g., this code from `tmol/score/dunbrack/potentials/potential.hh`
```
          // Perform the interpolation
	  TensorAccessor<Real, NbbP1 - 1, D> mean_slice(
	      rotameric_mean_tables.data() + tableind * rotameric_mean_tables.stride(0),
	      rotmean_table_sizes.data()->data()
	          + tableind * rotmean_table_sizes.stride(0),
	      rotmean_table_strides.data()->data()
	          + tableind * rotmean_table_strides.stride(0));
	
```
Creates a 2D slice of the 3-dimensional `rotameric_mean_tables` table. I pass in pointers to arrays of sizes and arrays of strides as the 2nd and 3rd arguments to TensorAccessor's constructor.

When combining a list of 2D tensors, into a single 3D tensor, if I have one tensor A of 10x10 and another tensor B of 20x20, then the joined 3D tensor, C, will be of size [2 x 20 x 20]. The data in the slice of [0, :, :] is put all right next to each other, so that
 * C[0, 0, 0] will hold the value of A[0,0],
 * C[0, 0, 9] will hold the value of A[0,9], and
 * C[0, 0, 10] will hold the value of A[1,0], etc.

Then the sizes tensor will be [ [ 10, 10 ], [20, 20] ] and the strides tensor will be [ [10, 1], [20, 1] ].