Add new scheme to support multiple independently-packed side chains
The idea is that each "sidechain" defines a part of the residue
to build from that is its "backbone" and it defines the root
of the side chain, the first side chain atom. Then a residue
can have multiple side chains, and each side chain will be
packed independently.

The idea with the backbone is that 1) the sidechain is where the
movement will be, but 2) the atom types of the backbone atoms
can change as the substituents change. In particular, the
atom type for Proline's backbone "N" changes from Nbb to
Npro when you mutate a non-pro residue to pro.

Not all atoms for a residue must be built when creating a
sidechain rotamer; however, no atom can appear in two
different sidechain rotamers on the same residue, or it
would be impossible to accurately calculate the energy
for them.