Add "final" function to script-module classes for blocking evaluation
The idea is that the forward() function (called principly via the __call__
method) returns an output tensor whose contents is not going to have
been written to necessarily until a synchronization event. The
"final" method invokes the function, and then synchronizes.
This function should be used in testing functions if you're not
passing through the score_component's "total" function
(which, because it uses the default stream, synchronizes).