Perform "slicing" of Kahan total within a 2nd kernel
So I spent yesterday wrestling with a weird "bug":

I had written a function to replace a series of if-thens
for determining what the largest power of two that was less
than an input value, x, was. I wrote it two different ways,
but both were similar in that when I printed out their
results, 99.8% of the time, I would get the right result,
and 0.2% of the time, I would get the wrong result.
There were only 32 possible input values.

It seems perhaps that printing out the result was unsable?

If I compare the result from the reference implementation,
and the result from my function, and only print out anything
if there's a discrepancy, then I get no output. This suggests
that the problem is in how the printf function in CUDA behaves.

Knowing printf is a problem and not __clz doesn't make me feel
much better.