Aleaverfay/stack elec (#219)
Stacked scoring for electrostatics

Interestingly, the electrostatics term brought to my attention a problem that I didn't know we had creatively sidestepped in Rosetta3: that problem of summing a large number of small values.

The previous (current) version of the electrostatics potential sums the interaction energies from many thousands of atom pairs. It uses `atomicAdd` so that each thread can write its contribution into the same total. The order in which threads complete is not deterministic and so the order the atomicAdds resolve is also non-deterministic. If you score the same structure 100 times in a row using this code, you get a WIDE variety of electrostatic energies, with a standard deviation of 6e-4 and a span of ~3e-4. This is probably enough fluctuation to prevent minimization from working well.

I tested Kahan summation which reduces the error in performing summations of a large number of values, but that proved to be incredibly slow.

Instead, I'm not accumulating the small values at double precision (they are still calculated at single precision) and then truncating from double precision to single precision in a final kernel that runs after the main kernel completes.

We should probably accumulate at double precision for the other atom-pair terms.