First working version of packing on the GPU.
Random number generation through curand using the
philox rng interface w/ NVidia.

NOTE: The more-recent versions of pytorch have changed
the interface for seed initialization and incrementing a
bit, so we will have to come back to this code (or whatever
version of the packer we end up using) when we update
to a newer release of pytorch.

This is a straight port of the CPU version, except
32 threads will perform the simulated annealing algorithm
in synchrony (with no inter-thread communication).

The result is SLOW. This is not an efficient version
of the packer. Gee wizz. I didn't time it, but it was
longer than a minute to run. I left the building and came
back maybe 10 minutes later and it had completed.