First pass alg for packing neighborhoods instead of entire structures
This is meant as a memory-savings technique because there isn't the
need to put enormous interaction graphs onto the GPU, which has limited
memory

So far, I merely have code for sectioning the IG into neighborhoods
and then invoking the packer repeatedly. It performs fine, but is slower.

If we are packing neighborhoods, then the reason to use the (slower)
chunk sparse matrix implementation disappears.