Preserve diversity in 1st round, refine in 2nd round
I want the pool of candidate backgrounds to stay wide
through the first round of neighborhood annealing, so
I have eliminated the quench step from the first round
and set the number of backgrounds to preserve to 300.
Also, the first round performs single-stage simA

Multistage simA is used in the second round.

This works decently well for the < 100 residue
proteins in my benchmark, but starts to be bad at
100 and 120. Larger proteins may need more rounds?