Fixed my pybind error
If a function pledges to return something
and doesn't, pybind will call that function a second time,
aparently? It's not clear. In any case, I forgot to make the
DunbrackChiSamplerTester::count_expanded_rotamers function
return the value that was returned by the
DunbrackChiSampler's version of the same function. In
all the flailing around trying to find my bug, I hadn't
even remembered that the DunbrackChiSamplerTest struct
even existed. ugh