Test obj2 offset for obj1 --> obj1 ind for obj2 mapper
There has to be a name for this: you have two collections
with thing A having several instances of thing B and you know
how many instances of thing B you have for each instance of
thing A, and can get a list of B-offsets for each A.  Having
this data, you then go and create an array where each entry
for thing B tells you which of thing A it belongs to.

E.g.

Three thing As with [3, 2, 3] instances of thing B each
so you get an offset array [0, 3, 5]
and then you can build the mapping
[ 0, 0, 0, 1, 1, 2, 2, 2]

I don't know the name for this operation, but I use it a lot.
I should figure out it's name, and move it to a utility function