Update numpy to 1.15+
This version offers access to a function, take_along_axis, which
is needed in branch aleaverfay/build_rotamers to sort the indices
of the rotamers in the input rotamer library by decreasing
probability.

Frustratingly, there is a bug in numpy above 1.14 that causes
calls to numpy.all and numpy.any which pass a torch tensor to
fail with the error:

```
TypeError: any() missing 1 required positional arguments: "dim"
```

as noted here: https://github.com/pytorch/pytorch/issues/15810

The workaround is to simply ask for the torch tensor as a numpy
array.