Add atoms-downstream-of-connection data to RefinedResidueType
Rename Residue in the datbase layer to RawResidueType
Rename ResidueType in the structure layer to RefinedResidueType

The atoms-downstream-of-connection data will help determine
the set of atoms that define a residue-spanning named DOF,
such as phi (which extends into the i-1 residue) or omega
(which extends two atoms deep into the i+1 residue). These
atoms are aluded to using the "UnresolvedAtom" in the
residue type parameter file (the chemical.yaml file),
and are turned into UnresolvedAtomIDs in the RefinedResidueType,
and with the inter-block connection data in the Pose, these
can be turned into PartialAtomIDs. The final bit of data
necessary is the identity of the atoms that are downstream
of the connection atoms, so that we can identify that
"CA" is the atom one bond deep into the i+1 residue for
the omega dihedral when bonded to an alpha amino acid;
if we were bonded to a beta amino acid, though, the
"CB" atom would be the 4th atom. If we want to support
alpha amino acids chemically bonded to beta amino acids,
we need this kind of indirection and atom-index resolution.