Add test for inverse + forward kinematics w/ rotamers
I can now retrieve the internal geometry from the
existing methionine rotamer at position 1 of 1ubq,
then copy the geometries for the mainchain atoms
of a leucine that I want to build at that position,
and assign values to chi1 and chi2. Then I invoke
the forward-folding step that computes coordinates,
but leaves them in their kintree order, so finally,
I reindex the coordinates.