Fix bug in identity xform
The homogenous coordinate for the homogenous transform's point
was set to 0 instead of 1 erroneously, meaning the coordinate
was not accumulated from the parent.

This is the kind of heizenbug that makes you scratch your head
and wonder how this code ever worked, or, at the very least,
what in particular caused it to fail in this case that was
different from all of the other cases. Truly baffling.

With this bugfix, the build_lots_of_rotamers unit test finally
works!