Add cudaEvent tracking to ensure that scoring doesn't race ahead
Torch, when it creates its streams, creates them to ignore the default
stream, so we cannot expect for energy calculation to stay in sync
with rotamer picking and coordinate copying into the alternate_coords
and alternate_id tensors by simply having that step occur in
the default stream. Similarly, we have to ask the default stream
to wait for scoring evaluation to complete in order to perform
the monte-carlo accept-reject step.