First attempt at score-term factory
All score term creators will have to live in tmol/score/terms
and will be automatically registered with the ScoreTermFactory
when someone (who??) types:

   from tmol.score.terms import *

using some file-discovery "magic" that I took from here:

https://stackoverflow.com/questions/1057431/how-to-load-all-modules-in-a-folder

The ScoreFunction class will create score terms when asked to set
a non-zero weight for a term. It will (but does not yet) pass the set of
"EnergyTermOptions" and "ChemicalDB" to the ScoreTermFactory when
it requests a new term so that the newly-created term can be initialized
with those parameters.