Resolve all previous uses of Pose and Poses
Following the creation of the PoseStack class, the code that previously
used the Pose and Poses class needed to be updated. This commit
touches quite a few files, but the changes are generally small.

The biggest differences come from code that previously interacted
with singular Pose objects, as a PoseSatck now has an extra dimension.
For this reason, some of the hard-coded tensor indices had to be increased
by 1.