Prune out missing connections from the set of torsion bonds
If the external atom in the center of a torsion points to
an inter-residue connection that is in fact unresolved
(something that really ought to never happen in a well
defined chemical system, but happens currently because we
don't have termini) then the neighbor residue (block) index
in the PoseStack for that inter-residue connection will
resolve to -1. So we have to prune such cases out.

I don't know if long term we ought to support unresolved
inter-residue connections because they are just weird.