Aleaverfay/minimize pose stack (#245)
This merge contains two main features:

1. Cartesian PoseStack minimization
2. Rapid construction of PoseStacks from sequence

The bulk of the work went into the second of these two bullets. A pose stack for 100 copies of ubiquitin can be constructed from sequence, scored, and have cartesian derivatives computed in in ~8ms on an A100.

The main interface for PoseStack construction lives in the PoseStackBuilder's `pose_stack_from_sequences` function and is intended to be the main way people will construct PoseStacks. This function takes

1. A PackedResidueTypes object, perhaps (encouraged to be) shared among multiple PoseStacks
2. A list of lists of amino acids / residue types by name, with optional annotations for non-polymeric connections (e.g. disulfides); 1-letter codes for the canonical L-AAs are accepted
3. A list of lists of chain lengths so that multiple chains for a single Pose in the stack can be given 

Each polymeric residue is understood to be chemically bound to the i-1 and i+1 residues (unless it's at either end of a chain). To designate non-polymeric connections, the syntax is to annotate the name of the residue type as follows:

```
    <resname>--<conn_name>-<conn_label>
```
where `<conn_name>` must be the name of one of the inter-residue connections for `<resname>` (e.g. `dslf` for `CYD`). The `<conn_label>` may be any string that the user wishes. Each `<conn_label>` should be specified exactly twice in a given structure so that the two residues which the user wants to signal are chemical bound both have the same `<conn_label>`. For example, a valid string for a 20-residue poly-alanine sequence with two disulfides might be:

```
    AA[CYD--dslf-first]AA[CYD--dslf-second]AAA[CYD--dslf-first]AAAAA[CYD--dslf-second]AAAA
                                               1               11111 1                1112
    12 3               45 6                789 0               12345 6                7890
```
(The ruler here on the 2nd and third lines are just to help ease of identifying residues by number).
Here, residue 3 is disulfide-bonded to residue 10 and residue 6 is disulfide-bonded to residue 16. The labels `first` and `second` could be replaced by any other.

During PoseStack construction, the inter-residue chemical bonds are fed into a GPU-friendly version of Warshall's algorithm to determine all pairs of inter-residue path distances between connection points; this algorithm scales poorly (O(N^3)) and future versions of this code will include a limited distance version of Djikstra's algorithm instead.