Rework CPU's DeviceOperations::foreach_workgroup
The idea is that the work for one chunk (e.g.
to evaluate the energy between two residue pairs) will contain
some sections where the CPU will run one for loop a single
time, e.g., to load data into a local array, and then
a second for loop, e.g  toiterate across lots of work.

The there are two sets of lambdas here: the outer lambda and
the inner lambda. The outer lambda will call the inner lambdas
once each; so we don't want the outer lambda to run several
iterations.