Add device-agnostic, residue-centric, whole-pose energy evaluation (#246)
Create a unified CPU/GPU implementation of the LJ/LK score term that still benefits from the use of shared memory. The same code is run on the CPU as the GPU except in the "dispatch" layer -- the use of a for loop on the CPU and a kernel launch on the GPU -- and the introduction of "intra-workgroup" communication in the form of synchronization and reductions (both no ops on the CPU). Data for a residue pair is loaded into shared memory so that work can be "tiled." The tiling code is just complex enough that it is worth putting into a templated function to avoid 1) duplicating the complex logic surrounding synchronization, and 2) getting the synchronization logic wrong.

100 copies of 1ubq can be scored in 3.4ms on the A100.

CPU performance of residue-centric scoring is not better than for pose-centric scoring; the slowdown is ~1.6x.

GPU performance of residue-centric scoring is better than pose-centric scoring; the speedup is ~2.2x.