Two things: code cleanup and also performance debugging that breaks scoring
The test here is: what happens if we don't spend
memory bandwidth loading water coordinates into
shared memory and we decrese our shared memory footprint
to 0 for water coords.

It turns out, this is not a massive savings; 1.4x speedup at 100
copies of 1ubq. Nothing to sneeze at, but not a 3x or 10x speedup.

It suggests that using less memory for water coords could be helpful

That requires dynamic allocation of shared memory. I like this task,
but it will be a challenge