First stab at an lk-ball that parallelizes across wat/occ and wat/wat intxns
I'm just roughing out the code at this point, it is still a long
way from compliing.

The plan from here is to the implement a fine-grained parallelization of lk-ball
for inter-residue forward calculations in parallel and get the algorithm working
on the CPU, then the GPU, then see if that changes 1) register use and 2) perf
and then move on to implement the backward pass, where I hope to see the
greatest improvement in performance.

I would like to standardize/generalize the code a little; it currently feels
very fragile and also duplicated between the wat/occ calc and the wat/wat calc.

Miles to go