Aleaverfay/res centric lkball (#249)
Add a residue-centric version of the LKBall term:
In the forward pass, the term launches one kernel to calculate the water coordinates for each of the polar atoms, and a second set of kernels for evaluating the residue-pair energies after. Residue coordinates are loaded into shared memory for both steps.
This term is slower than other terms relative to the CPU implementation, but significant effort in trying to improve performance was fruitless.