Implement disulfide score term (#254)
Implement res-centric disulfide score term.

We run a simple disulfide connection detection algorithm copied from Rosetta on the test PDB load to form the actual connections in the pose stack. The disulfide bonds are annotated in the term python code from the 'dslf' connections in the block type definitions. We flag all disulfide connections in a tensor and pass that into the C++ code, where it computes the potential for any valid flagged connections.