Aleaverfay/res centric rama (#252)
This implements the Ramachandran (rama_prepro) term in a residue-centric way. This is a two-body energy as which table from residue i depends on the amino acid identity of residue i+1. The scores match the baseline from R3. This PR starts using the custom uaid_t numpy.dtype for holding arrays of unresolved atom ids