Fix bug with strides of 0 for non-Scalar enable_tensor_views of tensors
torch.sum(x), e.g., returns a scalar in its backward-pass which
masquerades as a tensor with the same shape as x, except its strides
will give it away as not being a real tensor: they will all be 0.
But, if x is a tensor of coordinates and we use TCAST to coerce this
tensor into a TView<Eigen::Matrix<float, 3, 1>, N, D>, then the fact
that x and y and z all have the same address (because a stride of 0
means "ignore the index in the given dimension" and a tensor with
only strides of zeros means "ignore all the indices used to look
up data, there is only one value") will be LOST and the y and z
values will be read off the end of the array!

To fix this, we can simply look at any consumed dimension as determined
by the enable_tensor_view<SomeClass<T>> class and assert that the consumed
dimension(s) has (have) a non-zero stride. Any other stride besides zero
is fine.