Extending view_tensor stride checking to ensure contiguous allocation
Classes that define enable_tensor_view must consume contiguous blocks
of memory and cannot be allocated in a way where the strides of the
"consumed" dimensions represent anything but contiguous memory.

Therefore, the strides of consumed dimension i must be
   prod(j = i+1, j < input_tensor.dim(), input_tensor.size(j))

The strides for the remaining dimensions can be whatever they
need to be, but the consumed dimensions do need to be contiguous.

Imagine we have a tensor named "coords" which we are casting to
an Eigen::Matrix<float, 3, 1>. If coords is allocated
as a strange tensor of:
[n_poses x n_atoms x 3 x 15],
and we passed it to the c++ layer as
coords[:, :, :, 7]
then the "xyz" for a particular [i, j] would actually
read from
  x = coords[i, j, 0, 7] (correct)
  y = coords[i, j, 0, 8] (error)
  z = coords[i, j, 0, 9] (error)
because the Eigen::Matrix is going to calculate the locations for
y and z as being right next to x.

We can detect this issue by asserting that the stride for
the consumed dimension (dimension 2, the third dimension)
is exactly 1 and not 15, as it would be in this case.