Fix bug with strides of 0 for non-Scalar enable_tensor_views of tensors (#262)
torch.sum(x) in its backward-pass returns a *scalar*, 1, which masquerades as a derivative tensor with the same shape as x, except its strides will give it away as not being a real tensor: they will all be 0. But, if x is a tensor of coordinates and we use TCAST to coerce this derivative tensor into a TView<Eigen::Matrix<float, 3, 1>, N, D>, then the fact that x and y and z all have the same address (because a stride of 0 means "ignore the index in the given dimension" and a tensor with only strides of zeros means "ignore all the indices used to look up data, there is only one value") will be LOST and the y and z values will be read off the end of the array!

To fix this, we can simply look at any consumed dimension as determined by the enable_tensor_view<SomeClass<T>> class and assert that the consumed dimension(s) has (have) a non-zero stride. 

We should also check to ensure contiguous allocation of the memory that is consumed by the enable-tensor-view classes:

Classes that define enable_tensor_view must consume contiguous blocks of memory and cannot be allocated in a way where the strides of the "consumed" dimensions represent anything but contiguous memory.

Therefore, the strides of consumed dimension i must be
```
   prod(j = i+1, j < input_tensor.dim(), input_tensor.size(j))
```
The strides for the remaining dimensions can be whatever they need to be, but the consumed dimensions do need to be contiguous.

Imagine we have a tensor named "coords" which we are casting to an Eigen::Matrix<float, 3, 1>. If coords is allocated
as a strange tensor of:
```
   [n_poses x n_atoms x 3 x 15],
```
and we passed it to the c++ layer as
```
   coords[:, :, :, 7]
```
then the "xyz" for a particular [i, j] would actually
read from
```
  x = coords[i, j, 0, 7] // (correct)
  y = coords[i, j, 0, 8] // (error)
  z = coords[i, j, 0, 9] // (error)
```
because the Eigen::Matrix is going to calculate the locations for
y and z as being right next to x.

We can detect this issue by asserting that the stride for the consumed dimension (dimension 2, the third dimension) is exactly 1 and not 15, as it would be in this case.