Sketching code for improved dslf-detection logic
I want the user to be able to say "Here are the disulfides I know about"
and for the code to be smart enough to say "oh, look, there are some
other residues that might be disulfide-bonded. let's investigate."

My first thought had been that if the user provides a list of known
disulfides that no further disulfide detection would take place, but
it would be really unfortunate if an un-expected disulfide were
designed in that tmol then refused to acknowledge because it the
user had said "oh, I know about this disulfide." So instead of
closing the door on disulfide detection, when the user provides
a list of known disulfide connections, the code looks to see if there
are any additional CYS that are not described as being part of
a disulfide. If not, then it proceeds to run the (CPU-based)
disulfide-detection step, but ignores the CYS that the user has
indicated are forming disulfides.