Adding option to pose-building API for discontinuous segments
The optional argument res_not_connected for pose_stack_from_canonical_form
lets the user indicate that some residues should not be treated as termini
or as bonded to adjacent (prev and/or next) residues. The tensor is
   [n-poses x max-n-residues x 2]
where [:, :, 0] represents whether the residue should be treated as
not connected to the previous position, and where [:, :, 1] represents
whether the residue should be treated as not connected to the next residue.

If a residue is not to be treated as connected to either the previous or
next residue, then it will not be built as a termini variant either:
If a residue should be treated as a termini, then the residue should be
the first/last residue with a given chain id and that will be detected
using separate logic. In the case that the discontinuous segment is
not the n-terminal residue for a chain, but it is the first residue
for that chain, then res_not_connected[p, i, 0] serves to say "do not
treat this residue as an n-terminal residue" rather than to mean
"do not form a chemical bond to the previous residue" as that would
be implied already by the fact that the residue is not the same
chain as the previous residue. The same goes for the last residue
in a chain, where building an unwanted OXT might be especially disruptive.