Add atom_is_hydrogen data member to PackedBlockTypes
Structural rearrangement here:
The strings in RawResidueType/RefinedResidueType that refer to
the atom types and by extension to the elements are essentially
pointers that cannot be dereferenced in the absence of the
ChemicalDatabase that contains their properties and so
splitting a RefinedResidueType from the ChemicalDatabase that
defines it -- or at least that defines the properties it is
reliant on -- is a bad thing to do. To rememdy this, I am adding
references to the source ChemicalDatabase to both ResidueTypeSet
and PackedBlockTypes. This changes a lot of the unit tests responsible
for building PoseStacks and PackedBlockType objects; especially the
older tests that build from the old-style Residue objects