Aleaverfay/pose building interface (#257)
This PR creates an interface to tmol so that NN packages can output coordinates and residue-type information for run-of-the-mill proteins in a canonical form and then construct a PoseStack object needed for score function evaluation. There are a few features/drawbacks here worth mentioning:

1) the code as written can only handle the 20 canonical amino acids and the run-of-the-mill n- and c-termini variants,
2) chemical type resolution for HIS (HIS-E vs HIS-D) and CYS (CYS vs CYD) are handled as special cases,
3) if hydrogen atoms are not provided, their positions can be computed (details below)
4) the (greedy) disulfide chemical resolution steps is performed on the CPU

Details:
HIS chemical type resolution:
Histadine (at neutral pH) exists in one of two tautomerization states (90% of the time): 1) where the NE2 nitrogen is protonated and the ND1 nitrogen is not, and 2) where the ND1 nitrogen is protonated and the NE2 nitrogen is not. There are four possible ways that tmol will resolve this ambiguity:
   1) the user provides either the HD1 or the HE2 hydrogen coordinates (but not both) and tmol selects the corresponding tautomerization state
   2) the user provides a coordinate for "HN," a stand in for the hydrogen that is chemically bound to one of the HIS's two nitrogens; tmol will use the distance of HN to ND1 and to NE2 to and will declare the tautomerization state based on the shorter of the two
   3) the user provides the coordinates for three atoms "HN", "NH" and "NN" which are stand-ins for the hydrogen attached to the nitrogen, the nitrogen attached to the hydrogen, and the nitrogen that is not attached to the hydrogen. The distances between NH and NN and the ring's CG atom are used to decide which of the two tautomerization states the ring is in.
   4) the user provides only the heavy-atom coordinates, in which case the code chooses the NE2-is-protonated tautomerization state.

CYS disulfide detection:
Disulfides are detected by distance between SG atoms with a maximum detection distance of 2.5A. The algorithm is greedy: cysteine residues are iterated across in order in an outer loop (i) and again in an inner loop (j). The closest distance between SG on CYS i and any other as-of-yet-unpaired CYS j within 2.5A is found and a disulfide bond declared between i and j. This does not guarantee that the largest number of possible disulfides is found; if res 1 and 3 have a distance of 2.2A and 1 and 2 has 2.4A, and 3 and 4 have a distance of 2.4A, e.g. then the largest number of disulfides would be formed when 1 and 2 are paired and 3 and 4 are paired. This algorithm, however, will pair residues 1 and 3. Also of note: if the order of the residues changes, but their distances do not, so that 1 and 2 have a distance of 2.4A and then 2 and 3 have the distance of 2.2A, (and cysteines 3 and 4 continue to have a distance of 2.4A) and for sake of argument the other pairs have distances > 2.5, then the 1-2 disulfide will be found first, and therefore the 3-4 disulfide will also be found later.

If you already know what which pairs of CYS residues should be disulfide bonded, they can be given to the pose_stack_from_canonical_form call via a list of integer 3-tuples of (pose-ind, res1-ind, res2-ind); if there are no additional cysteins in the PoseStack besides those listed in this list, then the CPU-based step of finding disulfide pairs is skipped. Additionally, it is possible to skip the disulfide detection step entirely with the find_additional_disulfides flag.

H-placement:
If hydrogen atoms are missing, they will be built from ideal coordinates wrt heavy atoms. This is a fully-differentiable process so derivatives wrt the score function that apply to the hydrogen atoms get converted into derivatives for the atoms that define their coordinates. In the case of HIS, the NE2-protonated tautomerization state is chosen; in the case of SER, THR, CYS, and TYR, the hydroxyl's dihedral angle is assigned to 180. Future versions of this code will make more intelligent decisions on how to choose the dihedral angle, e.g. by looking for nearby hydrogen bond acceptors.

The logic for H-atom placement is extended to any heavy atom whose position can be known directly from the coordinates of other heavy atoms and who is a leaf atom in the icoor tree. Such atoms include O and OXT.

Inter-residue connectivity:
The chain identifier for each residue is given as input to the pose_stack_from_canonical_form interface. Residues i and i+1 from the same chain are presumed to form a chemical bond. The first and last residue of each chain is presumed to be a terminus. This behavior can be altered, however, using the res_not_connected argument: if residue i on pose p should not be presumed to form a chemical bond to the i-1 residue, then res_not_connected[p][i][0] should be set to True; if it should not be presumed to form a chemical bond to the i+1 residue, then res_not_connected[p][i][1] should be set to True. These residues will not be constructed with their respective termini variant patches applied. E.g. If the only residue in a chain (both the first and last residue) is marked as not connecting to the i-1 residue, but is not marked as not connecting to the i+1 residue, it will be built as having the C-terminal patch but not the N-terminal patch.

Gap residues are indicated in the input residue-type tensor by the sentinel index of -1. Such residues will be excised and all remaining residues in the given pose will be "left justified" so that they are adjacent in the residue-index ordering. That is, if residues i and i+j are part of the same chain and have only gap residues between them, then they will be presumed to form a polymeric chemical bond between them.